(incomplete) add optparse and allow program to run as a certain user
This makes the program a little more dynamic. More options on the
command line when starting up. Port can be specified in the config file
now, user doesn’t have to be the current user, and favicon is a
placeholder for now.

This change is not functioning completely at the moment, I wanted to
submit what I had to pick up later.